requirements.txt
python-telegram-bot==13.7
requests==2.22.0
openai==0.27.0
python-dotenv==1.0.0
psycopg2
APScheduler==3.10.4
pytz
python-dateutil
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
